leg 1 basic working
You can have leg 1 take a step and draw back manually in all directions.

+++ src/StepPlanner.h
@@ -34,7 +34,7 @@ class StepPlanner {
     void updateEndpoint(int16_t newControlCoordinateX, int16_t newControlCoordinateY);
     bool update();
 
-    int16_t getStepHeight(int16_t footXYDropL);
+    int16_t getStepHeight(float footXYDropL);
     void returnToOrigin();
     // bool footAtOrigin();
     // void reset();
